+++ magiccap/gui/index.html
@@ -258,6 +258,6 @@
                 </section>
             </div>
         </div>
-        <script src="gui_2.0.js"></script>
+        <script src="gui.js"></script>
     </body>
 </html>
